Move everything to jsx files, use capitals for component file names
