+++ src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './app.js';
+import App from './App';
 
 render(<App />, document.querySelector('#root'));
